Refatoração e adição de funcionalidades: remoção de código desnecessário, criação de modelo e migração para anúncios, e ajustes em requisições e componentes Vue.

+++ app/Http/Controllers/Api/ClienteController.php
@@ -19,35 +19,36 @@ class ClienteController extends BaseController
     public function index()
     {
 
-        $clientes = Cliente::with('bairro','user')->paginate(10);
-        return $this->sendResponse($clientes,'lista de clientes');
-
+        $clientes = Cliente::with('bairro', 'user')->paginate(10);
+        return $this->sendResponse($clientes, 'lista de clientes');
     }
-   public function all()
-   {
-       # code...
-       $clientes = Cliente::with('bairro','user')->get();
-       return $this->sendResponse($clientes,'lista de clientes');
-   }
+    public function all()
+    {
+        # code...
+        $clientes = Cliente::with('bairro', 'user')->get();
+        return $this->sendResponse($clientes, 'lista de clientes');
+    }
 
     public function store(Request $request)
     {
-       $this->validate($request,[ "name"=>"required|string|min:6|max:60",
-       "email" => "required|unique:users,email",
-       "password"=>"required|confirmed",
-       'bairro'=>"required"]);
+        $this->validate($request, [
+            "name" => "required|string|min:6|max:60",
+            "email" => "required|unique:users,email",
+            "password" => "required|confirmed",
+            'bairro' => "required"
+        ]);
 
-         DB::beginTransaction();
+        DB::beginTransaction();
 
         try {
             //code...
-        $bairro = Bairro::findOrFail($request->bairro);
-        $request->merge(["password" => Hash::make($request->password)]); // hash password
-        $user = User::create(['email'=>$request->email,'password'=>$request->password]);
-        $request->merge(['user_id',$user->id]);
-        $cliente  = $bairro->clientes()->create($request->except('bairro','password_confirmation','email','name','password'));
-        $funcao = Role::where('name','cliente')->first();
-        $user->assignRole([$funcao->id]);
+            $bairro = Bairro::findOrFail($request->bairro);
+            $request->merge(["password" => Hash::make($request->password)]); // hash password
+            $user = User::create(['email' => $request->email, 'password' => $request->password, 'name' => $request->name]);
+            $request->merge(['user_id' => $user->id, 'nome' => $request->name]);
+            $cliente  = $bairro->clientes()->create($request->except('bairro', 'password_confirmation', 'email', 'name', 'password'));
+            $funcao = Role::where('name', 'cliente')->first();
+            $user->assignRole([$funcao->id]);
         } catch (\Throwable $th) {
             //throw $th;
             DB::rollBack();
@@ -60,16 +61,18 @@ class ClienteController extends BaseController
 
     public function show($id)
     {
-        $cliente = Cliente::with('bairro','pedidos')->findOrFail($id);
-        return $this->sendResponse($cliente,'cliente');
+        $cliente = Cliente::with('bairro', 'pedidos')->findOrFail($id);
+        return $this->sendResponse($cliente, 'cliente');
     }
 
     public function update(Request $request, $id)
     {
-        $this->validate($request,[ "name"=>"required|string|min:6|max:60",
-        "email" => "required",
-        "password"=>"confirmed",
-        'bairro'=>"required"]);
+        $this->validate($request, [
+            "name" => "required|string|min:6|max:60",
+            "email" => "required",
+            "password" => "confirmed",
+            'bairro' => "required"
+        ]);
 
         DB::beginTransaction();
 
@@ -79,22 +82,20 @@ class ClienteController extends BaseController
                 # code...
                 $request->merge(["password" => Hash::make($request->password)]);
             } // hash password
-         $cliente = Cliente::findOrFail($id);
-         if (isset($request->password)) {
-             # code...
-           $cliente->user()->update(['name'=>$request->name,'email'=>$request->email,"password"=>$request->password]);
-         } else {
-             # code...
-              $cliente->user()->update(['name'=>$request->name,'email'=>$request->email]);
-         }
+            $cliente = Cliente::findOrFail($id);
+            if (isset($request->password)) {
+                # code...
+                $cliente->user()->update(['name' => $request->name, 'email' => $request->email, "password" => $request->password]);
+            } else {
+                # code...
+                $cliente->user()->update(['name' => $request->name, 'email' => $request->email]);
+            }
 
 
-         $request->merge(['nome'=>$request->name,'bairro_id'=>$request->bairro,'user_id'=>$cliente->user_id]);
-         $cliente->updateOrCreate($request->except('bairro','password_confirmation','email','name','password'));
-         $funcao = Role::where('name','cliente')->first();
-         ($cliente->user->hasRole('cliente')) ?   $cliente->user->syncRoles([$funcao->id]) :  $cliente->user->assignRole([$funcao->id]) ;
-
-
+            $request->merge(['nome' => $request->name, 'bairro_id' => $request->bairro, 'user_id' => $cliente->user_id]);
+            $cliente->updateOrCreate($request->except('bairro', 'password_confirmation', 'email', 'name', 'password'));
+            $funcao = Role::where('name', 'cliente')->first();
+            ($cliente->user->hasRole('cliente')) ?   $cliente->user->syncRoles([$funcao->id]) :  $cliente->user->assignRole([$funcao->id]);
         } catch (\Throwable $th) {
             //throw $th;
             DB::rollBack();
@@ -102,28 +103,28 @@ class ClienteController extends BaseController
         }
         DB::commit();
         return $this->sendResponse($cliente, 'Cliente actualizado com sucesso.');
-
-
     }
 
 
     public function destroy($id)
     {
-          DB::beginTransaction();
+        DB::beginTransaction();
         try {
-             //code...
-             $record = Cliente::findOrFail($id);
-             $record->delete();
-         } catch (\Throwable $th) {
-             DB::rollBack();
-             return $this->sendError('Erro ao eliminar os dados', $th->getMessage());
-         }
+            //code...
+            $record = Cliente::findOrFail($id);
+            //se o cliente  tiver um pedido nao podemos deletar o cliente e lançar uma exceção
+            if (count($record->pedidos) > 0) {
 
-         DB::commit();
-         return $this->sendResponse($record, 'Cliente eliminado com sucesso.');
+                throw new \Exception("Não é possivel eliminar o cliente, pois o cliente tem pedidos associados.");
+            }
 
+            $record->delete();
+        } catch (\Throwable $th) {
+            DB::rollBack();
+            return $this->sendError('Erro ao eliminar os dados', $th->getMessage(),403);
+        }
+
+        DB::commit();
+        return $this->sendResponse($record, 'Cliente eliminado com sucesso.');
     }
-
-
-
 }
